added viewfragment functionality.
