+++ app/src/main/AndroidManifest.xml
@@ -26,10 +26,10 @@
             android:name=".MainActivity"
             android:exported="false"/>
         <activity
-            android:name=".Register"
+            android:name=".auth.Register"
             android:exported="false"/>
         <activity
-            android:name=".LoginPage"
+            android:name=".auth.LoginPage"
             android:exported="true">
