--- app/src/main/java/com/example/fishbook/Map.java
@@ -1,6 +1,0 @@
-package com.example.fishbook;
-
-import androidx.fragment.app.Fragment;
-
-public class Map extends Fragment {
-}
