--- app/src/main/java/com/example/fishbook/MapActivity.java
@@ -1,16 +1,0 @@
-package com.example.fishbook;
-
-import android.os.Bundle;
-
-import androidx.appcompat.app.AppCompatActivity;
-import androidx.fragment.app.Fragment;
-
-public class MapActivity extends AppCompatActivity {
-    @Override
-    public void onCreate(Bundle savedInstanceState){
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_map);
-        Fragment fragment = new MapFragment();
-        getSupportFragmentManager().beginTransaction().replace(R.id.frame_layout, fragment).commit();
-    }
-}
