--- app/src/main/java/com/example/fishbook/Register.kt
@@ -1,75 +1,0 @@
-package com.example.fishbook
-
-import android.content.Intent
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-import android.widget.Button
-import android.widget.EditText
-import android.widget.TextView
-import android.widget.Toast
-
-import com.google.firebase.auth.FirebaseAuth
-import com.google.firebase.auth.ktx.auth
-import com.google.firebase.ktx.Firebase
-
-class Register : AppCompatActivity() {
-    lateinit var etEmail: EditText
-    private lateinit var etPass: EditText
-    private lateinit var btnSignUp: Button
-    lateinit var textView: TextView
-
-
-    private lateinit var auth: FirebaseAuth
-
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_register)
-
-
-        etEmail = findViewById(R.id.email)
-
-        etPass = findViewById(R.id.password)
-        btnSignUp = findViewById(R.id.btn_register)
-        textView = findViewById(R.id.loginNow)
-
-
-        auth = Firebase.auth
-
-        btnSignUp.setOnClickListener {
-            signUpUser()
-        }
-
-
-        textView.setOnClickListener {
-            val intent = Intent(this, LoginPage::class.java)
-            startActivity(intent)
-        }
-
-    }
-
-    private fun signUpUser() {
-        val email = etEmail.text.toString()
-        val pass = etPass.text.toString()
-
-
-
-        if (email.isBlank() || pass.isBlank()) {
-            Toast.makeText(this, "Email and Password can't be blank", Toast.LENGTH_SHORT).show()
-            return
-        }
-
-
-        auth.createUserWithEmailAndPassword(email, pass).addOnCompleteListener(this) {
-            if (it.isSuccessful) {
-                Toast.makeText(this, "Successfully Singed Up", Toast.LENGTH_SHORT).show()
-                val intent = Intent(this, LoginPage::class.java)
-                startActivity(intent)
-                finish()
-            } else {
-                Toast.makeText(this, "Singed Up Failed!", Toast.LENGTH_SHORT).show()
-            }
-        }
-    }
-}
