+++ app/src/main/java/com/example/fishbook/gallery/GalleryViewModel.kt
@@ -2,6 +2,8 @@ package com.example.fishbook.gallery
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
+import com.example.fishbook.record.CatchDetails
+import com.google.firebase.auth.FirebaseAuth
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.asStateFlow
@@ -9,15 +11,18 @@ import kotlinx.coroutines.launch
 
 class GalleryViewModel : ViewModel() {
     private val imageRepository = ImageRepository.get()
+    val userId = FirebaseAuth.getInstance().currentUser?.uid
 
-    private val _GridImages: MutableStateFlow<List<GridImage>> = MutableStateFlow(emptyList())
-    val GridImages: StateFlow<List<GridImage>>
-        get() = _GridImages.asStateFlow()
+    private val _catchDetails: MutableStateFlow<List<CatchDetails>> = MutableStateFlow(emptyList())
+    val CatchDetails: StateFlow<List<CatchDetails>>
+        get() = _catchDetails.asStateFlow()
 
     init {
         viewModelScope.launch {
-            imageRepository.fetchImages().collect {
-                _GridImages.value = it
+            if (userId != null) {
+                imageRepository.fetchImages(userId).collect {
+                    _catchDetails.value = it
+                }
             }
         }
     }
