--- app/src/main/java/com/example/fishbook/gallery/GridImage.kt
@@ -1,3 +1,0 @@
-package com.example.fishbook.gallery
-
-data class GridImage(val id: String = "", val url: String = "")
