+++ app/src/main/java/com/example/fishbook/gallery/ViewRecordFragment.kt
@@ -10,6 +10,7 @@ import com.example.fishbook.databinding.FragmentViewRecordBinding
 import com.squareup.picasso.Picasso
 import androidx.navigation.fragment.navArgs
 import com.example.fishbook.gallery.ViewRecordFragmentArgs
+import com.example.fishbook.record.CatchDetails
 
 
 class ViewRecordFragment : Fragment() {
@@ -27,8 +28,8 @@ class ViewRecordFragment : Fragment() {
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
-        val gridImage = args.gridImage
-        updateUi(gridImage)
+        val catchDetail = args.catchDetail
+        updateUi(catchDetail)
     }
 
     override fun onDestroyView() {
@@ -36,9 +37,17 @@ class ViewRecordFragment : Fragment() {
         _binding = null
     }
 
-    private fun updateUi(gridImage: GridImage) {
+    private fun updateUi(catchDetail: CatchDetails) {
         binding.apply {
-            Picasso.get().load(gridImage.url).into(largeImage)
+            Picasso.get().load(catchDetail.remoteUri).into(largeImage)
+           tvSpecies.text = catchDetail.species
+            tvCounty.text = catchDetail.county
+            tvLake.text = catchDetail.lake
+            tvLure.text = catchDetail.lure
+
+            tvWeight.text = catchDetail.weight.toString()
+            tvLength.text = catchDetail.length.toString()
+
         }
     }
 }
