+++ app/src/main/java/com/example/fishbook/map/Map.java
@@ -1,0 +1,6 @@
+package com.example.fishbook.map;
+
+import androidx.fragment.app.Fragment;
+
+public class Map extends Fragment {
+}
