+++ app/src/main/java/com/example/fishbook/map/MapActivity.java
@@ -1,0 +1,19 @@
+package com.example.fishbook.map;
+
+import android.os.Bundle;
+
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.fragment.app.Fragment;
+
+import com.example.fishbook.R;
+import com.example.fishbook.map.MapFragment;
+
+public class MapActivity extends AppCompatActivity {
+    @Override
+    public void onCreate(Bundle savedInstanceState){
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_map);
+        Fragment fragment = new MapFragment();
+        getSupportFragmentManager().beginTransaction().replace(R.id.frame_layout, fragment).commit();
+    }
+}
