+++ app/src/main/res/layout/activity_map.xml
@@ -4,7 +4,7 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MapActivity"
+    tools:context=".map.MapActivity"
     android:orientation="vertical"
     android:layout_margin="20dp">
