--- app/src/main/res/layout/fragment_camera.xml
@@ -1,113 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<ScrollView xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".Camera">
-
-    <RelativeLayout
-        android:layout_width="match_parent"
-        android:layout_height="match_parent">
-
-        <ImageView
-            android:id="@+id/fish_image"
-            android:layout_width="match_parent"
-            android:layout_height="400dp"
-            android:scaleType="centerCrop"
-            android:background="@color/black"/>
-
-        <ImageButton
-            android:id="@+id/camera_button"
-            android:layout_width="48dp"
-            android:layout_height="48dp"
-            android:layout_margin="10dp"
-            android:layout_below="@id/fish_image"
-            android:layout_centerHorizontal="true"
-            android:src="@drawable/ic_camera" />
-
-        <EditText
-            android:id="@+id/species_edit_text"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_margin="10dp"
-            android:layout_below="@id/camera_button"
-            android:hint="@string/species_hint"
-            android:inputType="text" />
-
-        <EditText
-            android:id="@+id/lure_edit_text"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_margin="10dp"
-            android:layout_below="@id/camera_button"
-            android:layout_centerHorizontal="true"
-            android:hint="@string/lure_hint"
-            android:inputType="text"/>
-
-        <EditText
-            android:id="@+id/lake_edit_text"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_margin="10dp"
-            android:layout_below="@id/species_edit_text"
-            android:hint="@string/lake_hint"
-            android:inputType="text"/>
-
-        <EditText
-            android:id="@+id/county_edit_text"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_margin="10dp"
-            android:layout_below="@id/lure_edit_text"
-            android:layout_centerHorizontal="true"
-            android:hint="@string/county_hint"
-            android:inputType="text"/>
-
-        <EditText
-            android:id="@+id/length_edit_text"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_margin="10dp"
-            android:layout_below="@id/county_edit_text"
-            android:hint="@string/length_hint"
-            android:inputType="numberDecimal"/>
-
-        <EditText
-            android:id="@+id/weight_edit_text"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_margin="10dp"
-            android:layout_below="@id/lake_edit_text"
-            android:layout_centerHorizontal="true"
-            android:hint="@string/weight_hint"
-            android:inputType="numberDecimal"/>
-
-        <EditText
-            android:id="@+id/time_edit_text"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_margin="10dp"
-            android:layout_below="@id/length_edit_text"
-            android:hint="@string/time_hint"
-            android:inputType="datetime"/>
-
-        <EditText
-            android:id="@+id/location_edit_text"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_margin="10dp"
-            android:layout_below="@id/weight_edit_text"
-            android:layout_centerHorizontal="true"
-            android:hint="@string/location_hint"
-            android:inputType="text"/>
-
-        <Button
-            android:id="@+id/upload_button"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_below="@id/location_edit_text"
-            android:layout_centerHorizontal="true"
-            android:layout_margin="10dp"
-            android:text="@string/upload"/>
-    </RelativeLayout>
-</ScrollView>
